Added alternate Set value pattern for Blackboard Enum Keys
Now that we show all of the enum keys if no EnumClass is selected, you can select the EnumKey before an EnumClass is selected.

I added a code path for this case, so that it will fill-in the EnumClass based on the EnumKey that was selected from the Blackboard.

The original sequencing still works (selecting EnumClass then EnumKey), but having both workflows makes it more intuitive and less error-prone.